Add SwiftUI MMVM-like design pattern, separating model from view code.
SwiftUI is prone to have massive view code. This way, we can separate view code from business logic. The view handles rendering and events, fetching its information from the model, and forwarding events to the model. Quite similar to MMVM. Makes use of an extension to separate the view and the model into separate files.